requirements.txt
requests
pillow
shutil
discord